refactor(cli): extract the preview-bundle format into :bundle-format
The `.previewbundle` format — its manifest DTO, well-known entry names,
sidecar injectors, deterministic zip helpers, detached signature scheme,
classpath hydration, Android resource/launch support and figma-raster
bound — was a `:cli` internal. It is an external, versioned format read by
the daemon, the viewer and the Gradle plugin, and #3824 preparation item 5
asks for it to become a module of its own.

It is one now: `:bundle-format` (`bundle/format`), depending only on
`:common-io` and `:preview-data-api`. The build enforces that — nothing in
it can reach `:cli`, `serve`, the daemon, or a data product. The `bundle`
subcommands stay in `:cli`; they are argument parsing, not format.

Two things did not move. `readWebEmbedData` returns `WebEmbed.Preview`, and
`WebEmbed` reaches into `serve`'s escaping helpers, so it stays in `:cli` as
`readBundleWebEmbedData` — a bundle reader should not need an HTML generator
on its classpath. And the types keep their `ee.schimke.composeai.cli`
package for source-compat, as `:gradle-preview-driver` did for its own
carve-out, so every existing call site resolves unchanged.

That package choice has a consequence worth stating plainly: `check-serve-
seam.py` keys a crossing on the package an import names, so all eleven
bundle entries stay on `cliInternalsUsedByServe`. The register is not stale
— an extracted server still cannot name an unpublished module, whatever
build it sits in. Finishing item 5 means renaming the package to
`ee.schimke.composeai.bundle` and publishing, which forces
`unpublishedContracts` off empty until the next release publishes the
module. That is a release-sequencing decision, so it gets its own PR rather
than riding along with a file move.

Also fixes two path filters that would have gone quietly stale:
`serve-lanes-paths.json` and `serve-lanes-e2e.yml` pin
`AndroidBundleResources.kt` / `AndroidBundleLaunch.kt` by path, and both now
point at the new location.
