+++ bundle/format/src/test/kotlin/ee/schimke/composeai/cli/BundleClasspathHydrationTest.kt
