+++ scripts/serve-seam-allowlist.json
@@ -9,6 +9,11 @@
     "server would have to get from a published contract module instead. The bundle-format cluster",
     "(BundleReader, BundleSigning, BundleClasspathHydration, extractBundle*, locateBundleSidecar*,",
     "BUNDLE_VERSION, ZIP_DOS_EPOCH_MS) is #3824 preparation item 5 and is most of this list.",
+  "Those symbols now live in their own module, `:bundle-format`, but kept the",
+  "`ee.schimke.composeai.cli` package for source-compat, and this check keys a crossing on the",
+  "package an import names rather than the module that declares it. So they are still listed,",
+  "and correctly: an extracted server cannot name an unpublished module, whatever build it sits",
+  "in. They come off when item 5 finishes \u2014 package rename plus publishing.",
     "serveInternalsUsedByCli: the CLI reaching into serve. Each entry must either move to the",
     "server build with serve, or move down into a contract module the CLI can keep using. The",
     "long tail here is ServeCommand.kt, which #3824 wants reduced to a thin entry point.",
